microservicio/pyproject.toml
[project]
name = "microservicio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cachetools>=5.0",
    "fastapi>=0.136.1",
    "numpy>=2.4.4",
    "openai>=1.0.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=5.4.1",
    "uvicorn[standard]>=0.46.0",
]
